+++ akamai/resource_akamai_dns_record.go
@@ -4,20 +4,21 @@ import (
 	"encoding/hex"
 	"encoding/json"
 	"fmt"
+	dnsv2 "github.com/akamai/AkamaiOPEN-edgegrid-golang/configdns-v2"
+	"github.com/hashicorp/terraform/helper/schema"
+	"github.com/hashicorp/terraform/helper/validation"
 	"log"
 	"net"
+	"regexp"
 	"sort"
 	"strconv"
 	"strings"
 	"sync"
-
-	dnsv2 "github.com/akamai/AkamaiOPEN-edgegrid-golang/configdns-v2"
-	"github.com/hashicorp/terraform/helper/schema"
-	"github.com/hashicorp/terraform/helper/validation"
+	"time"
 )
 
 // Retry count for save, update and delete
-const opRetryCount = 3
+const opRetryCount = 5
 
 func resourceDNSv2Record() *schema.Resource {
 	return &schema.Resource{
@@ -69,6 +70,9 @@ func resourceDNSv2Record() *schema.Resource {
 					RRTypeSoa,
 					RRTypeAkamaiCdn,
 					RRTypeAkamaiTlc,
+					RRTypeCaa,
+					RRTypeCert,
+					RRTypeTlsa,
 				}, false),
 			},
 			"ttl": {
@@ -80,11 +84,10 @@ func resourceDNSv2Record() *schema.Resource {
 				Optional: true,
 			},
 			"target": {
-				Type:             schema.TypeSet,
+				Type:             schema.TypeList,
 				Elem:             &schema.Schema{Type: schema.TypeString},
 				Optional:         true,
 				DiffSuppressFunc: dnsRecordTargetSuppress,
-				Set:              schema.HashString,
 			},
 			"subtype": {
 				Type:     schema.TypeInt,
@@ -278,12 +281,97 @@ func resourceDNSv2Record() *schema.Resource {
 				Type:     schema.TypeInt,
 				Optional: true,
 			},
+			"usage": {
+				Type:     schema.TypeInt,
+				Optional: true,
+			},
+			"selector": {
+				Type:     schema.TypeInt,
+				Optional: true,
+			},
+			"match_type": {
+				Type:     schema.TypeInt,
+				Optional: true,
+			},
+			"certificate": {
+				Type:     schema.TypeString,
+				Optional: true,
+			},
+			"type_value": {
+				Type:             schema.TypeInt,
+				Optional:         true,
+				DiffSuppressFunc: dnsRecordTypeValueSuppress,
+			},
+			"type_mnemonic": {
+				Type:     schema.TypeString,
+				Optional: true,
+			},
 			"record_sha": {
 				Type:     schema.TypeString,
 				Computed: true,
 			},
 		},
 	}
+}
+
+/*
+https://tools.ietf.org/html/rfc4398 section 2.1 defines cert types
+
+             0            Reserved
+             1  PKIX      X.509 as per PKIX
+             2  SPKI      SPKI certificate
+             3  PGP       OpenPGP packet
+             4  IPKIX     The URL of an X.509 data object
+             5  ISPKI     The URL of an SPKI certificate
+             6  IPGP      The fingerprint and URL of an OpenPGP packet
+             7  ACPKIX    Attribute Certificate
+             8  IACPKIX   The URL of an Attribute Certificate
+         9-252            Available for IANA assignment
+           253  URI       URI private
+           254  OID       OID private
+*/
+
+var certTypes = map[string]int{
+	"PKIX":    1,
+	"SPKI":    2,
+	"PGP":     3,
+	"IPKIX":   4,
+	"ISPKI":   5,
+	"IPGP":    6,
+	"ACPKIX":  7,
+	"IACPKIX": 8,
+	"URI":     253,
+	"OID":     254,
+}
+
+// Suppress check for type_value. Mnemonic config comes back as numeric
+func dnsRecordTypeValueSuppress(k, old, new string, d *schema.ResourceData) bool {
+
+	oldv, newv := d.GetChange("type_value")
+	oldval := oldv.(int)
+	newval := newv.(int)
+	typeval := 0
+	mnemonictype := d.Get("type_mnemonic").(string)
+	mnemonicvalue, ok := certTypes[mnemonictype]
+	if !ok {
+		return false
+	}
+	if oldval == 0 && newval == 0 {
+		return true
+	} else if oldval != 0 && newval != 0 {
+		typeval = oldval
+	} else if oldval == 0 {
+		typeval = newval
+	} else if newval == 0 {
+		typeval = oldval
+	} else {
+		return false
+	}
+	if typeval == mnemonicvalue {
+		return true
+	}
+
+	return false
 }
 
 // DiffSuppresFunc to handle quoted TXT Rdata strings possibly containining escaped quotes
@@ -294,15 +382,14 @@ func dnsRecordTargetSuppress(k, old, new string, d *schema.ResourceData) bool {
 	// Seems if different, get one invocation with old val and null new as well as
 	// invocation with new val and null old. In all cases, we retrieve old and new sets
 	// from ResourceData and validate thru those.
-	// Function only validates TXT records.
 
-	oldset, newset := d.GetChange("target")
-	oldTargetList := oldset.(*schema.Set).List()
-	newTargetList := newset.(*schema.Set).List()
+	recordtype := d.Get("recordtype").(string)
+	oldlist, newlist := d.GetChange("target")
+	oldTargetList := oldlist.([]interface{})
+	newTargetList := newlist.([]interface{})
 	if len(oldTargetList) != len(newTargetList) {
 		return false
 	}
-	//var baseList []interface{}
 	var compList []interface{}
 	var baseVal = ""
 	var compTrim = ""
@@ -314,13 +401,11 @@ func dnsRecordTargetSuppress(k, old, new string, d *schema.ResourceData) bool {
 		if strings.Contains(baseVal, "\\\"") {
 			baseVal = strings.ReplaceAll(baseVal, "\\\"", "\"")
 		}
-		//baseList = oldTargetList
 		compList = newTargetList
 	} else if old == "" {
 		baseVal = new
 		compTrim = "\\\""
 		baseVal = strings.Trim(baseVal, "\"")
-		//baseList = newTargetList
 		compList = oldTargetList
 	} else if new == "" {
 		baseVal = old
@@ -329,9 +414,24 @@ func dnsRecordTargetSuppress(k, old, new string, d *schema.ResourceData) bool {
 		if strings.Contains(baseVal, "\\\"") {
 			baseVal = strings.ReplaceAll(baseVal, "\\\"", "\"")
 		}
-		//baseList = oldTargetList
 		compList = newTargetList
 	}
+
+	if recordtype == "CAA" {
+		basevalsplit := strings.Split(strings.Trim(baseVal, "\""), "\"")
+		baseVal = strings.Join(basevalsplit, "")
+		for _, compval := range compList {
+			compvalsplit := strings.Split(strings.Trim(compval.(string), "\""), "\"")
+			compval = strings.Join(compvalsplit, "")
+			log.Printf("updated baseVal: %v", baseVal)
+			log.Printf("compval: %v", compval)
+			if baseVal == strings.Trim(compval.(string), "\"") {
+				return true
+			}
+		}
+		return false
+	}
+
 	for _, compval := range compList {
 		if compTrim == "\\\"" && strings.Contains(compval.(string), "\\\"") {
 			compval = strings.ReplaceAll(compval.(string), "\\\"", "\"")
@@ -353,6 +453,7 @@ var recordCreateLock = map[string]*sync.Mutex{
 	"AKAMAICDN":  &sync.Mutex{},
 	"AKAMAITLC":  &sync.Mutex{},
 	"CAA":        &sync.Mutex{},
+	"CERT":       &sync.Mutex{},
 	"CNAME":      &sync.Mutex{},
 	"HINFO":      &sync.Mutex{},
 	"LOC":        &sync.Mutex{},
@@ -386,6 +487,7 @@ var recordTypeMapLock = &sync.Map{
         "AKAMAICDN":    &sync.Map{},
         "AKAMAITLC":    &sync.Map{},
         "CAA":          &sync.Map{},
+        "CERT":         &sync.Map{},
         "CNAME":        &sync.Map{},
         "HINFO":        &sync.Map{},
         "LOC":          &sync.Map{},
@@ -431,12 +533,14 @@ func executeRecordFunction(name string, d *schema.ResourceData, fn recordFunctio
 
 	// DNS API can have Concurrency issues
 	opRetry := opRetryCount
-	e := fn(zone, false)
+	e := fn(zone, rlock)
 	for e != nil && opRetry > 0 {
 		if dnsv2.IsConfigDNSError(e) {
-			if e.(*dnsv2.RecordError).ConcurrencyConflict() {
+			if e.(dnsv2.ConfigDNSError).ConcurrencyConflict() {
+				log.Printf("[WARNING] [Akamai DNSv2] Concurrency Conflict")
 				opRetry -= 1
-				e = fn(zone, false)
+				time.Sleep(100 * time.Millisecond)
+				e = fn(zone, rlock)
 				continue
 			} else if name == "DELETE" && e.(dnsv2.ConfigDNSError).NotFound() == true {
 				// record doesn't exist
@@ -467,6 +571,8 @@ func resourceDNSRecordCreate(d *schema.ResourceData, meta interface{}) error {
 	var zone string
 	var host string
 	var recordtype string
+
+	log.Printf("[INFO] [Akamai DNS] Record Create")
 
 	_, ok := d.GetOk("zone")
 	if ok {
@@ -503,10 +609,13 @@ func resourceDNSRecordCreate(d *schema.ResourceData, meta interface{}) error {
 	log.Printf("[DEBUG] [Akamai DNSv2] SHA sum for recordcreate [%s]", sha1hash)
 	// First try to get the zone from the API
 	log.Printf("[DEBUG] [Akamai DNSv2] Searching for records [%s]", zone)
-
-	rdata, e := dnsv2.GetRdata(zone, host, recordtype)
+	rdata := make([]string, 0, 0)
+	recordset, e := dnsv2.GetRecord(zone, host, recordtype)
 	if e != nil && !dnsv2.IsConfigDNSError(e) {
 		return fmt.Errorf("error looking up "+recordtype+" records for %q: %s", host, e)
+	}
+	if recordset != nil {
+		rdata = dnsv2.ProcessRdata(recordset.Target, recordtype)
 	}
 	// If there's no existing record we'll create a blank one
 	if dnsv2.IsConfigDNSError(e) && e.(dnsv2.ConfigDNSError).NotFound() == true {
@@ -553,6 +662,8 @@ func resourceDNSRecordUpdate(d *schema.ResourceData, meta interface{}) error {
 	// this prevents lost data if you are using a counter/dynamic variables
 	// in your config.tf which might overwrite each other
 
+	log.Printf("[INFO] [Akamai DNS] Record Update")
+
 	var zone string
 	var host string
 	var recordtype string
@@ -571,7 +682,7 @@ func resourceDNSRecordUpdate(d *schema.ResourceData, meta interface{}) error {
 	}
 	_, ok = d.GetOk("target")
 	if ok {
-		target := d.Get("target").(*schema.Set).List()
+		target := d.Get("target").([]interface{})
 		records := make([]string, 0, len(target))
 		for _, recContent := range target {
 			records = append(records, recContent.(string))
@@ -617,12 +728,14 @@ func resourceDNSRecordUpdate(d *schema.ResourceData, meta interface{}) error {
 	log.Printf("[DEBUG] [Akamai DNSv2] UPDATE SHA sum for recordupdate [%s]", sha1hash)
 	// First try to get the zone from the API
 	log.Printf("[DEBUG] [Akamai DNSv2] UPDATE Searching for records [%s]", zone)
-
-	rdata, e := dnsv2.GetRdata(zone, host, recordtype)
+	rdata := make([]string, 0, 0)
+	recordset, e := dnsv2.GetRecord(zone, host, recordtype)
 	if e != nil && !dnsv2.IsConfigDNSError(e) {
 		return fmt.Errorf("error looking up "+recordtype+" records for %q: %s", host, e)
 	}
-
+	if recordset != nil {
+		rdata = dnsv2.ProcessRdata(recordset.Target, recordtype)
+	}
 	log.Printf("[DEBUG] [Akamai DNSv2] UPDATE Searching for records LEN %d", len(rdata))
 	if len(rdata) > 0 {
 		sort.Strings(rdata)
@@ -666,6 +779,8 @@ func resourceDNSRecordRead(d *schema.ResourceData, meta interface{}) error {
 	var host string
 	var recordtype string
 
+	log.Printf("[INFO] [Akamai DNS] Record Read")
+
 	_, ok := d.GetOk("zone")
 	if ok {
 		zone = d.Get("zone").(string)
@@ -689,7 +804,7 @@ func resourceDNSRecordRead(d *schema.ResourceData, meta interface{}) error {
 	}
 
 	log.Printf("[DEBUG] [Akamai DNSv2] READ record JSON from bind records %s %s %s %s", string(b), zone, host, recordtype)
-	//sort.Strings(recordcreate.Target)
+	sort.Strings(recordcreate.Target)
 	extractString := strings.Join(recordcreate.Target, " ")
 	sha1hash := getSHAString(extractString)
 	log.Printf("[DEBUG] [Akamai DNSv2] READ SHA sum for Existing SHA test %s %s", extractString, sha1hash)
@@ -719,7 +834,31 @@ func resourceDNSRecordRead(d *schema.ResourceData, meta interface{}) error {
 	}
 	log.Printf("[DEBUG] [Akamai DNSv2] READ record data read JSON %s", string(b1))
 	rdataFieldMap := dnsv2.ParseRData(recordtype, record.Target) // returns map[string]interface{}
-	if recordtype != "MX" {
+	if recordtype == "MX" {
+		// calc rdata sha from read record
+		sort.Strings(record.Target)
+		rdataString := strings.Join(record.Target, " ")
+		shaRdata := getSHAString(rdataString)
+		if d.HasChange("target") {
+			log.Printf("MX READ. TARGET HAS CHANGED")
+			// has remote changed independently of TF?
+			if d.Get("record_sha").(string) != shaRdata {
+				return fmt.Errorf("Recordset [%s %s]: Remote has diverged from TF Config. Manual intervention required.", host, recordtype)
+			} else {
+				log.Printf("MX READ. Remote static")
+				d.SetId("")
+			}
+		} else {
+			log.Printf("MX READ. TARGET HAS NOT CHANGED")
+			// has remote changed independently of TF?
+			if d.Get("record_sha").(string) != shaRdata {
+				// another special case ... for instances record sha might not be representative of full resource
+				if len(d.Get("target").([]interface{})) != 1 || sha1hash != shaRdata {
+					return fmt.Errorf("Recordset [%s %s]: Remote has diverged from TF Config. Manual intervention required.", host, recordtype)
+				}
+			}
+		}
+	} else {
 		// Parse Rdata. MX special
 		for fname, fvalue := range rdataFieldMap {
 			d.Set(fname, fvalue)
@@ -819,12 +958,17 @@ func resourceDNSRecordImport(d *schema.ResourceData, meta interface{}) ([]*schem
 	d.Set("name", recordset.Name)
 	d.Set("recordtype", recordset.RecordType)
 	d.Set("ttl", recordset.TTL)
-	// Parse Rdata
-	rdataFieldMap := dnsv2.ParseRData(recordset.RecordType, recordset.Target) // returns map[string]interface{}
-	for fname, fvalue := range rdataFieldMap {
-		d.Set(fname, fvalue)
-	}
 	targets := dnsv2.ProcessRdata(recordset.Target, recordtype)
+	if recordset.RecordType == "MX" {
+		// can't guarantee order of MX records. Forced to set pri, incr to 0 and targets as is
+		d.Set("target", targets)
+	} else {
+		// Parse Rdata
+		rdataFieldMap := dnsv2.ParseRData(recordset.RecordType, recordset.Target) // returns map[string]interface{}
+		for fname, fvalue := range rdataFieldMap {
+			d.Set(fname, fvalue)
+		}
+	}
 	importTargetString := ""
 	if len(targets) > 0 {
 		if recordtype != "MX" {
@@ -845,6 +989,8 @@ func resourceDNSRecordImport(d *schema.ResourceData, meta interface{}) ([]*schem
 
 func resourceDNSRecordDelete(d *schema.ResourceData, meta interface{}) error {
 
+	log.Printf("[INFO] [Akamai DNS] Record Delete")
+
 	zone := d.Get("zone").(string)
 	host := d.Get("name").(string)
 	recordtype := d.Get("recordtype").(string)
@@ -854,7 +1000,7 @@ func resourceDNSRecordDelete(d *schema.ResourceData, meta interface{}) error {
 	getRecordLock(zone, host, recordtype).Lock()
 	defer getRecordLock(zone, host, recordtype).Unlock()
 
-	target := d.Get("target").(*schema.Set).List()
+	target := d.Get("target").([]interface{})
 
 	records := make([]string, 0, len(target))
 	for _, recContent := range target {
@@ -873,6 +1019,8 @@ func resourceDNSRecordExists(d *schema.ResourceData, meta interface{}) (bool, er
 	var zone string
 	var host string
 	var recordtype string
+
+	log.Printf("[INFO] [Akamai DNS] Record Exists")
 
 	_, ok := d.GetOk("zone")
 	if ok {
@@ -974,30 +1122,30 @@ func bindRecord(d *schema.ResourceData) (dnsv2.RecordBody, error) {
 	}
 
 	ttl := d.Get("ttl").(int)
-	target := d.Get("target").(*schema.Set).List()
+	target := d.Get("target").([]interface{})
 	records := make([]string, 0, len(target))
 
-	simplerecordtarget := map[string]bool{"AAAA": true, "CNAME": true, "LOC": true, "NS": true, "PTR": true, "SPF": true, "SRV": true, "TXT": true}
+	simplerecordtarget := map[string]bool{"AAAA": true, "CNAME": true, "LOC": true, "NS": true, "PTR": true, "SPF": true, "SRV": true, "TXT": true, "CAA": true}
 
 	for _, recContent := range target {
 		if simplerecordtarget[recordtype] {
-
-			if recordtype == "AAAA" {
+			switch recordtype {
+			case "AAAA":
 				addr := net.ParseIP(recContent.(string))
 				result := FullIPv6(addr)
 				log.Printf("[DEBUG] [Akamai DNSv2] IPV6 full %s", result)
 				records = append(records, result)
-			} else if recordtype == "LOC" {
+			case "LOC":
 				log.Printf("[DEBUG] [Akamai DNSv2] LOC code format %s", recContent.(string))
 				str := padCoordinates(recContent.(string))
 				records = append(records, str)
-			} else if recordtype == "SPF" {
+			case "SPF":
 				str := recContent.(string)
 				if !strings.HasPrefix(str, "\"") {
 					str = "\"" + str + "\""
 				}
 				records = append(records, str)
-			} else if recordtype == "TXT" {
+			case "TXT":
 				str := recContent.(string)
 				log.Printf("[DEBUG] [Akamai DNSv2] Bind TXT Data IN: [%s]", str)
 				if strings.HasPrefix(str, "\"") {
@@ -1018,7 +1166,12 @@ func bindRecord(d *schema.ResourceData) (dnsv2.RecordBody, error) {
 				}
 				log.Printf("[DEBUG] [Akamai DNSv2] Bind TXT Data OUT: [%s]", str)
 				records = append(records, str)
-			} else {
+			case "CAA":
+				caaparts := strings.Split(recContent.(string), " ")
+				caaparts[2] = strings.Trim(caaparts[2], "\"")
+				caaparts[2] = "\"" + caaparts[2] + "\""
+				records = append(records, strings.Join(caaparts, " "))
+			default:
 				checktarget := recContent.(string)[len(recContent.(string))-1:]
 				if checktarget == "." {
 					records = append(records, recContent.(string))
@@ -1033,7 +1186,7 @@ func bindRecord(d *schema.ResourceData) (dnsv2.RecordBody, error) {
 
 	emptyrecordcreate := dnsv2.RecordBody{}
 
-	simplerecord := map[string]bool{"A": true, "AAAA": true, "AKAMAICDN": true, "CNAME": true, "LOC": true, "NS": true, "PTR": true, "SPF": true, "TXT": true}
+	simplerecord := map[string]bool{"A": true, "AAAA": true, "AKAMAICDN": true, "CNAME": true, "LOC": true, "NS": true, "PTR": true, "SPF": true, "TXT": true, "CAA": true}
 	if simplerecord[recordtype] {
 		sort.Strings(records)
 
@@ -1122,19 +1275,21 @@ func bindRecord(d *schema.ResourceData) (dnsv2.RecordBody, error) {
 			zone := d.Get("zone").(string)
 
 			log.Printf("[DEBUG] [Akamai DNSv2] MX record targets to process: %v", target)
-
-			rdata, e := dnsv2.GetRdata(zone, host, recordtype)
+			recordset, e := dnsv2.GetRecord(zone, host, recordtype)
+			rdata := make([]string, 0, 0)
 			if e != nil {
-				log.Printf("[DEBUG] [Akamai DNSv2] Searching for existing MX records no prexisting targets found LEN %d", len(rdata))
+				if !dnsv2.IsConfigDNSError(e) || !e.(dnsv2.ConfigDNSError).NotFound() {
+					// failure other than not found
+					return dnsv2.RecordBody{}, fmt.Errorf(e.Error())
+				} else {
+					log.Printf("[DEBUG] [Akamai DNSv2] Searching for existing MX records no prexisting targets found")
+				}
+			} else {
+				rdata = dnsv2.ProcessRdata(recordset.Target, recordtype)
 			}
 			log.Printf("[DEBUG] [Akamai DNSv2] Existing MX records to append to target %v", rdata)
 
-			records := make([]string, 0, len(target)+len(rdata))
-
-			priority := d.Get("priority").(int)
-			increment := d.Get("priority_increment").(int)
-			log.Printf("[DEBUG] [Akamai DNSv2] MX BIND Priority: %d ; Increment: %d", priority, increment)
-
+			//create map from rdata
 			rdataTargetMap := make(map[string]int, len(rdata))
 			for _, r := range rdata {
 				entryparts := strings.Split(r, " ")
@@ -1144,34 +1299,74 @@ func bindRecord(d *schema.ResourceData) (dnsv2.RecordBody, error) {
 				}
 				rdataTargetMap[rn], _ = strconv.Atoi(entryparts[0])
 			}
+			if d.HasChange("target") {
+				// see if any entry was deleted. If so, remove from rdata map.
+				oldlist, newlist := d.GetChange("target")
+				oldTargetList := oldlist.([]interface{})
+				newTargetList := newlist.([]interface{})
+				for _, oldtarg := range oldTargetList {
+					for _, newtarg := range newTargetList {
+						if oldtarg.(string) == newtarg.(string) {
+							break
+						}
+					}
+					// not there. remove
+					log.Printf("[DEBUG] [Akamai DNSv2] MX BIND target %v deleted", oldtarg)
+					deltarg := oldtarg.(string)
+					rdtparts := strings.Split(oldtarg.(string), " ")
+					if len(rdtparts) > 1 {
+						deltarg = rdtparts[1]
+					}
+					delete(rdataTargetMap, deltarg)
+				}
+			}
+			records := make([]string, 0, len(target)+len(rdata))
 
+			priority := d.Get("priority").(int)
+			increment := d.Get("priority_increment").(int)
+			log.Printf("[DEBUG] [Akamai DNSv2] MX BIND Priority: %d ; Increment: %d", priority, increment)
+			// walk thru target first
 			for _, recContent := range target {
 				targentry := recContent.(string)
 				if targentry[len(recContent.(string))-1:] != "." {
 					targentry += "."
 				}
-				if pri, ok := rdataTargetMap[targentry]; ok {
+				log.Printf("[DEBUG] [Akamai DNSv2] MX BIND Processing Target %s", targentry)
+				targhost := targentry
+				targpri := 0
+				targparts := strings.Split(targentry, " ") // need to support target entry with/without priority
+				if len(targparts) > 1 {
+					if len(targparts) > 2 {
+						return dnsv2.RecordBody{}, fmt.Errorf("Invalid MX Record format")
+					}
+					targhost = targparts[1]
+					var err error
+					targpri, err = strconv.Atoi(targparts[0])
+					if err != nil {
+						return dnsv2.RecordBody{}, fmt.Errorf("Invalid MX Record format")
+					}
+				} else {
+					targpri = priority
+				}
+				if pri, ok := rdataTargetMap[targhost]; ok {
 					log.Printf("MX BIND. %s in existing map", targentry)
 					// target already in rdata
-					if pri != priority {
+					if pri != targpri {
 						return dnsv2.RecordBody{}, fmt.Errorf("MX Record Priority Mismatch. Target order must align with EdgeDNS")
-					} else {
-						// already in rdata
-						continue
 					}
-				} else if containsPriority(rdataTargetMap, priority) {
-					log.Printf("MX BIND. %d already in RDATA", priority)
-					// already an rdata record with same priority
-					return dnsv2.RecordBody{}, fmt.Errorf("MX Record Priority Duplication. Target order and priority must align with EdgeDNS")
+					delete(rdataTargetMap, targhost)
 				}
-				records = append(records, strconv.Itoa(priority)+" "+targentry)
-
+				if len(targparts) == 1 {
+					records = append(records, strconv.Itoa(priority)+" "+targentry)
+				} else {
+					records = append(records, targentry)
+				}
 				if increment > 0 {
 					priority = priority + increment
 				}
 			}
 			log.Printf("[DEBUG] [Akamai DNSv2] Appended new target to target array LEN %d %v", len(records), records)
-
+			// append what ever is left ...
 			for targname, tpri := range rdataTargetMap {
 				records = append(records, strconv.Itoa(tpri)+" "+targname)
 			}
@@ -1329,6 +1524,35 @@ func bindRecord(d *schema.ResourceData) (dnsv2.RecordBody, error) {
 			recordcreate := dnsv2.RecordBody{Name: host, RecordType: recordtype, TTL: ttl, Target: records}
 			return recordcreate, nil
 		}
+		if recordtype == "CERT" {
+
+			records := make([]string, 0, len(target))
+			certtype := d.Get("type_mnemonic").(string)
+			typevalue := d.Get("type_value").(int)
+			keytag := d.Get("keytag").(int)
+			algorithm := d.Get("algorithm").(int)
+			certificate := d.Get("certificate").(string)
+			// value or mnemonic type?
+			if certtype == "" {
+				certtype = strconv.Itoa(typevalue)
+			}
+			records = append(records, certtype+" "+strconv.Itoa(keytag)+" "+strconv.Itoa(algorithm)+" "+certificate)
+
+			recordcreate := dnsv2.RecordBody{Name: host, RecordType: recordtype, TTL: ttl, Target: records}
+			return recordcreate, nil
+		}
+		if recordtype == "TLSA" {
+
+			records := make([]string, 0, len(target))
+			usage := d.Get("usage").(int)
+			selector := d.Get("selector").(int)
+			matchtype := d.Get("match_type").(int)
+			certificate := d.Get("certificate").(string)
+			records = append(records, strconv.Itoa(usage)+" "+strconv.Itoa(selector)+" "+strconv.Itoa(matchtype)+" "+certificate)
+
+			recordcreate := dnsv2.RecordBody{Name: host, RecordType: recordtype, TTL: ttl, Target: records}
+			return recordcreate, nil
+		}
 	}
 	return emptyrecordcreate, fmt.Errorf("Unable to create a Record Body for %s : %s", host, recordtype)
 
@@ -1374,6 +1598,12 @@ func validateRecord(d *schema.ResourceData) error {
 		return checkAkamaiTlcRecord(d)
 	case RRTypeSoa:
 		return checkSoaRecord(d)
+	case RRTypeCaa:
+		return checkCaaRecord(d)
+	case RRTypeCert:
+		return checkCertRecord(d)
+	case RRTypeTlsa:
+		return checkTlsaRecord(d)
 	default:
 		return fmt.Errorf("Invalid recordtype %v", recordtype)
 	}
@@ -1385,22 +1615,22 @@ func checkBasicRecordTypes(d *schema.ResourceData) error {
 	ttl := d.Get("ttl").(int)
 
 	if host == "" {
-		return fmt.Errorf("Type host must be set")
+		return fmt.Errorf("Configuration argument host must be set")
 	}
 
 	if recordtype == "" {
-		return fmt.Errorf("Type recordtype must be set")
+		return fmt.Errorf("Configuration argument recordtype must be set")
 	}
 
 	if ttl == 0 {
-		return fmt.Errorf("Type ttl must be set")
+		return fmt.Errorf("Configuration argument ttl must be set")
 	}
 
 	return nil
 }
 
 func checkTargets(d *schema.ResourceData) error {
-	target := d.Get("target").(*schema.Set).List()
+	target := d.Get("target").([]interface{})
 	records := make([]string, 0, len(target))
 
 	for _, recContent := range target {
@@ -1408,7 +1638,7 @@ func checkTargets(d *schema.ResourceData) error {
 	}
 
 	if len(records) == 0 {
-		return fmt.Errorf("Type records must be set.")
+		return fmt.Errorf("Configuration argument target must be set.")
 	}
 
 	return nil
@@ -1429,7 +1659,7 @@ func checkSimpleRecord(d *schema.ResourceData) error {
 func checkAsdfRecord(d *schema.ResourceData) error {
 	subtype := d.Get("subtype").(int)
 	if subtype == 0 {
-		return fmt.Errorf("Type subtype must be set for ASDF.")
+		return fmt.Errorf("Configuration argument subtype must be set for ASDF.")
 	}
 
 	if err := checkTargets(d); err != nil {
@@ -1447,23 +1677,23 @@ func checkDnskeyRecord(d *schema.ResourceData) error {
 	ttl := d.Get("ttl").(int)
 
 	if !(flags == 0 || flags == 256 || flags == 257) {
-		return fmt.Errorf("Type flags must not be %v for DNSKEY.", flags)
+		return fmt.Errorf("Configuration argument flags must not be %v for DNSKEY.", flags)
 	}
 
 	if ttl == 0 {
-		return fmt.Errorf("Type ttl must be set for DNSKEY.")
+		return fmt.Errorf("Configuration argument ttl must be set for DNSKEY.")
 	}
 
 	if protocol == 0 {
-		return fmt.Errorf("Type protocol must be set for DNSKEY.")
+		return fmt.Errorf("Configuration argument protocol must be set for DNSKEY.")
 	}
 
 	if !((algorithm >= 1 && algorithm <= 8) || algorithm != 10) {
-		return fmt.Errorf("Type algorithm must not be %v for DNSKEY.", algorithm)
+		return fmt.Errorf("Configuration argument algorithm must not be %v for DNSKEY.", algorithm)
 	}
 
 	if key == "" {
-		return fmt.Errorf("Type key must be set for DNSKEY.")
+		return fmt.Errorf("Configuration argument key must be set for DNSKEY.")
 	}
 
 	return nil
@@ -1476,19 +1706,19 @@ func checkDsRecord(d *schema.ResourceData) error {
 	digest := d.Get("digest").(string)
 
 	if digestType == 0 {
-		return fmt.Errorf("Type digest_type must be set for DS.")
+		return fmt.Errorf("Configuration argument digest_type must be set for DS.")
 	}
 
 	if keytag == 0 {
-		return fmt.Errorf("Type keytag must be set for DS.")
+		return fmt.Errorf("Configuration argument keytag must be set for DS.")
 	}
 
 	if algorithm == 0 {
-		return fmt.Errorf("Type algorithm must be set for DS.")
+		return fmt.Errorf("Configuration argument algorithm must be set for DS.")
 	}
 
 	if digest == "" {
-		return fmt.Errorf("Type digest must be set for DS.")
+		return fmt.Errorf("Configuration argument digest must be set for DS.")
 	}
 
 	return nil
@@ -1499,11 +1729,11 @@ func checkHinfoRecord(d *schema.ResourceData) error {
 	software := d.Get("software").(string)
 
 	if hardware == "" {
-		return fmt.Errorf("Type hardware must be set for HINFO.")
+		return fmt.Errorf("Configuration argument hardware must be set for HINFO.")
 	}
 
 	if software == "" {
-		return fmt.Errorf("Type software must be set for HINFO.")
+		return fmt.Errorf("Configuration argument software must be set for HINFO.")
 	}
 
 	return nil
@@ -1517,7 +1747,7 @@ func checkMxRecord(d *schema.ResourceData) error {
 	}
 
 	if priority < 0 || priority > 65535 {
-		return fmt.Errorf("Type priority must be set for MX.")
+		return fmt.Errorf("Configuration argument priority must be set for MX.")
 	}
 
 	if err := checkTargets(d); err != nil {
@@ -1540,27 +1770,27 @@ func checkNaptrRecord(d *schema.ResourceData) error {
 	}
 
 	if flagsnaptr == "" {
-		return fmt.Errorf("Type flagsnaptr must be set for NAPTR.")
+		return fmt.Errorf("Configuration argument flagsnaptr must be set for NAPTR.")
 	}
 
 	if order < 0 || order > 65535 {
-		return fmt.Errorf("Type order must not be %v for NAPTR.", order)
+		return fmt.Errorf("Configuration argument order must not be %v for NAPTR.", order)
 	}
 
 	if preference == 0 {
-		return fmt.Errorf("Type preference must be set for NAPTR.")
+		return fmt.Errorf("Configuration argument preference must be set for NAPTR.")
 	}
 
 	if regexp == "" {
-		return fmt.Errorf("Type regexp must be set for NAPTR.")
+		return fmt.Errorf("Configuration argument regexp must be set for NAPTR.")
 	}
 
 	if replacement == "" {
-		return fmt.Errorf("Type replacement must be set for NAPTR.")
+		return fmt.Errorf("Configuration argument replacement must be set for NAPTR.")
 	}
 
 	if service == "" {
-		return fmt.Errorf("Type service must be set for NAPTR.")
+		return fmt.Errorf("Configuration argument service must be set for NAPTR.")
 	}
 
 	return nil
@@ -1579,23 +1809,23 @@ func checkNsec3Record(d *schema.ResourceData) error {
 	}
 
 	if !(flags == 0 || flags == 1) {
-		return fmt.Errorf("Type flags must be set for NSEC3.")
+		return fmt.Errorf("Configuration argument flags must be set for NSEC3.")
 	}
 
 	if algorithm != 1 {
-		return fmt.Errorf("Type flags must be set for NSEC3.")
+		return fmt.Errorf("Configuration argument flags must be set for NSEC3.")
 	}
 	if iterations == 0 {
-		return fmt.Errorf("Type iterations must be set for NSEC3.")
+		return fmt.Errorf("Configuration argument iterations must be set for NSEC3.")
 	}
 	if nextHashedOwnerName == "" {
-		return fmt.Errorf("Type nextHashedOwnerName must be set for NSEC3.")
+		return fmt.Errorf("Configuration argument nextHashedOwnerName must be set for NSEC3.")
 	}
 	if salt == "" {
-		return fmt.Errorf("Type salt must be set for NSEC3.")
+		return fmt.Errorf("Configuration argument salt must be set for NSEC3.")
 	}
 	if typeBitmaps == "" {
-		return fmt.Errorf("Type typeBitMaps must be set for NSEC3.")
+		return fmt.Errorf("Configuration argument typeBitMaps must be set for NSEC3.")
 	}
 	return nil
 }
@@ -1611,19 +1841,19 @@ func checkNsec3ParamRecord(d *schema.ResourceData) error {
 	}
 
 	if !(flags == 0 || flags == 1) {
-		return fmt.Errorf("Type flags must be set for NSEC3PARAM.")
+		return fmt.Errorf("Configuration argument flags must be set for NSEC3PARAM.")
 	}
 
 	if algorithm != 1 {
-		return fmt.Errorf("Type algorithm must be set for NSEC3PARAM.")
+		return fmt.Errorf("Configuration argument algorithm must be set for NSEC3PARAM.")
 	}
 
 	if iterations == 0 {
-		return fmt.Errorf("Type iterations must be set for NSEC3PARAM.")
+		return fmt.Errorf("Configuration argument iterations must be set for NSEC3PARAM.")
 	}
 
 	if salt == "" {
-		return fmt.Errorf("Type salt must be set for NSEC3PARAM.")
+		return fmt.Errorf("Configuration argument salt must be set for NSEC3PARAM.")
 	}
 
 	return nil
@@ -1638,11 +1868,11 @@ func checkRpRecord(d *schema.ResourceData) error {
 	}
 
 	if mailbox == "" {
-		return fmt.Errorf("Type mailbox must be set for RP.")
+		return fmt.Errorf("Configuration argument mailbox must be set for RP.")
 	}
 
 	if txt == "" {
-		return fmt.Errorf("Type txt must be set for RP.")
+		return fmt.Errorf("Configuration argument txt must be set for RP.")
 	}
 
 	return nil
@@ -1664,40 +1894,39 @@ func checkRrsigRecord(d *schema.ResourceData) error {
 	}
 
 	if expiration == "" {
-		return fmt.Errorf("Type expiration must be set for RRSIG.")
+		return fmt.Errorf("Configuration argument expiration must be set for RRSIG.")
 	}
 
 	if inception == "" {
-		return fmt.Errorf("Type inception must be set for RRSIG.")
+		return fmt.Errorf("Configuration argument inception must be set for RRSIG.")
 	}
 
 	if originalTTL == 0 {
-		return fmt.Errorf("Type originalTTL must be set for RRSIG.")
+		return fmt.Errorf("Configuration argument originalTTL must be set for RRSIG.")
 	}
 
 	if algorithm == 0 {
-		return fmt.Errorf("Type algorithm must be set for RRSIG.")
+		return fmt.Errorf("Configuration argument algorithm must be set for RRSIG.")
 	}
 
 	if labels == 0 {
-
-		return fmt.Errorf("Type labels must be set for RRSIG.")
+		return fmt.Errorf("Configuration argument labels must be set for RRSIG.")
 	}
 
 	if keytag == 0 {
-		return fmt.Errorf("Type keytag must be set for RRSIG.")
+		return fmt.Errorf("Configuration argument keytag must be set for RRSIG.")
 	}
 
 	if signature == "" {
-		return fmt.Errorf("Type signature must be set for RRSIG.")
+		return fmt.Errorf("Configuration argument signature must be set for RRSIG.")
 	}
 
 	if signer == "" {
-		return fmt.Errorf("Type signer must be set for RRSIG.")
+		return fmt.Errorf("Configuration argument signer must be set for RRSIG.")
 	}
 
 	if typeCovered == "" {
-		return fmt.Errorf("Type typeCovered must be set for RRSIG.")
+		return fmt.Errorf("Configuration argument typeCovered must be set for RRSIG.")
 	}
 
 	return nil
@@ -1717,15 +1946,15 @@ func checkSrvRecord(d *schema.ResourceData) error {
 	}
 
 	if priority == 0 {
-		return fmt.Errorf("Type priority must be set for SRV.")
+		return fmt.Errorf("Configuration argument priority must be set for SRV.")
 	}
 
 	if weight < 0 || weight > 65535 {
-		return fmt.Errorf("Type weight must not be %v for SRV.", weight)
+		return fmt.Errorf("Configuration argument weight must not be %v for SRV.", weight)
 	}
 
 	if port == 0 {
-		return fmt.Errorf("Type port must be set for SRV.")
+		return fmt.Errorf("Configuration argument port must be set for SRV.")
 	}
 
 	return nil
@@ -1741,15 +1970,15 @@ func checkSshfpRecord(d *schema.ResourceData) error {
 	}
 
 	if algorithm == 0 {
-		return fmt.Errorf("Type algorithm must be set for SSHFP.")
+		return fmt.Errorf("Configuration argument algorithm must be set for SSHFP.")
 	}
 
 	if fingerprintType == 0 {
-		return fmt.Errorf("Type fingerprintType must be set for SSHFP.")
+		return fmt.Errorf("Configuration argument fingerprintType must be set for SSHFP.")
 	}
 
 	if fingerprint == "null" {
-		return fmt.Errorf("Type fingerprint must be set for SSHFP.")
+		return fmt.Errorf("Configuration argument fingerprint must be set for SSHFP.")
 	}
 
 	return nil
@@ -1759,38 +1988,33 @@ func checkSoaRecord(d *schema.ResourceData) error {
 
 	nameserver := d.Get("name_server").(string)
 	emailaddr := d.Get("email_address").(string)
-	serial := d.Get("serial").(int)
 	refresh := d.Get("refresh").(int)
 	retry := d.Get("retry").(int)
 	expiry := d.Get("expiry").(int)
 	nxdomainttl := d.Get("nxdomain_ttl").(int)
 
 	if nameserver == "" {
-		return fmt.Errorf("Key %s must be specified in SOA record", nameserver)
+		return fmt.Errorf("Configuration argument %s must be specified in SOA record", "nameserver")
 	}
 
 	if emailaddr == "" {
-		return fmt.Errorf("Key %s must be specified in SOA record", emailaddr)
-	}
-
-	if serial == 0 {
-		return fmt.Errorf("Key %d must be specified in SOA record", serial)
+		return fmt.Errorf("Configuration argument %s must be specified in SOA record", "emailaddr")
 	}
 
 	if refresh == 0 {
-		return fmt.Errorf("Key %d must be specified in SOA record", refresh)
+		return fmt.Errorf("Configuration argument %s must be specified in SOA record", "refresh")
 	}
 
 	if retry == 0 {
-		return fmt.Errorf("Key %d must be specified in SOA record", retry)
+		return fmt.Errorf("Configuration argument %s must be specified in SOA record", "retry")
 	}
 
 	if expiry == 0 {
-		return fmt.Errorf("Key %d must be specified in SOA record", expiry)
+		return fmt.Errorf("Configuration argument %s must be specified in SOA record", "expiry")
 	}
 
 	if nxdomainttl == 0 {
-		return fmt.Errorf("Key %d must be specified in SOA record", nxdomainttl)
+		return fmt.Errorf("Configuration argument %s must be specified in SOA record", "nxdomainttl")
 	}
 
 	return nil
@@ -1801,14 +2025,102 @@ func checkAkamaiTlcRecord(d *schema.ResourceData) error {
 	answertype := d.Get("answer_type").(string)
 
 	if dnsname != "" {
-		return fmt.Errorf("dnsname key is computed. It must not be set in AKAMAITLC.")
+		return fmt.Errorf("Configuration argument dnsname is computed. It must not be set in AKAMAITLC.")
 	}
 
 	if answertype != "" {
-		return fmt.Errorf("answertype key is computed. It must not be set in AKAMAITLC.")
+		return fmt.Errorf("Configuration argument answertype is computed. It must not be set in AKAMAITLC.")
 	}
 
 	return nil
+}
+
+func checkCaaRecord(d *schema.ResourceData) error {
+
+	if err := checkBasicRecordTypes(d); err != nil {
+		return err
+	}
+
+	if err := checkTargets(d); err != nil {
+		return err
+	}
+
+	caatarget := d.Get("target").([]interface{})
+	for _, caa := range caatarget {
+		caaparts := strings.Split(caa.(string), " ")
+		if len(caaparts) != 3 {
+			return fmt.Errorf("Configuration argument CAA target %s is invalid.", caa.(string))
+		}
+
+		flag, err := strconv.Atoi(caaparts[0])
+		if err != nil || flag < 0 || flag > 255 {
+			return fmt.Errorf("Configuration argument CAA target %s is invalid. flag value must be <= 0 and >= 255.", caa.(string))
+		}
+		re := regexp.MustCompile(`[^a-zA-Z0-9]+`)
+		submatchall := re.FindAllString(caaparts[1], -1)
+		if len(submatchall) > 0 {
+			return fmt.Errorf("Configuration argument  CAA target %s is invalid. tag contains invalid characters.", caa.(string))
+		}
+	}
+
+	return nil
+}
+
+func checkCertRecord(d *schema.ResourceData) error {
+
+	typemnemonic := d.Get("type_mnemonic").(string)
+	typevalue := d.Get("type_value").(int)
+	//keytag := d.Get("keytag").(int)
+	//algorithm := d.Get("algorithm").(int)
+	certificate := d.Get("certificate").(string)
+
+	if err := checkBasicRecordTypes(d); err != nil {
+		return err
+	}
+
+	if typemnemonic == "" && typevalue == 0 {
+		return fmt.Errorf("Configuration arguments type_value and type_mnemonic are not set. Invalid CERT configuration.")
+	}
+
+	if typemnemonic != "" && typevalue != 0 {
+		return fmt.Errorf("Configuration arguments type_value and type_mnemonic are both set. Invalid CERT configuration.")
+	}
+
+	if certificate == "" {
+		return fmt.Errorf("Configuration argument certificate must be set for CERT.")
+	}
+	/*
+	   if algorithm == 0 {
+	           return fmt.Errorf("Configuration argument algorithm must be set for CERT.")
+	   }
+
+	   if keytag == 0 {
+	           return fmt.Errorf("Configuration argument keytag must be set for CERT.")
+	   }
+	*/
+	return nil
+
+}
+
+func checkTlsaRecord(d *schema.ResourceData) error {
+
+	usage := d.Get("usage").(int)
+	certificate := d.Get("certificate").(string)
+
+	if err := checkBasicRecordTypes(d); err != nil {
+		return err
+	}
+
+	if certificate == "" {
+		return fmt.Errorf("Configuration argument certificate must be set for TLSA.")
+	}
+
+	if usage == 0 {
+		return fmt.Errorf("Configuration argument usage must be set for TLSA.")
+	}
+
+	return nil
+
 }
 
 // Resource record types supported by the Akamai Edge DNS API
@@ -1838,4 +2150,5 @@ const (
 	RRTypeNsec3      = "NSEC3"
 	RRTypeNsec3Param = "NSEC3PARAM"
 	RRTypeRrsig      = "RRSIG"
+	RRTypeCert       = "CERT"
 )