+++ examples/akamai_dns_zone/add-zone-records/dnsv2.tf
@@ -1,68 +1,67 @@
 provider "akamai" {
-    edgerc = "/root/.edgerc"
-    dns_section = "dns"
+  dns_section = "dns"
 }
 
 locals {
   zone = "akavaiodeveloper.net"
 }
 
-resource "akamai_dnsv2_zone" "test_zone" {
-    contractid = "C-1FRYVV3"
-    zone = "${local.zone}"
+resource "akamai_dns_zone" "test_zone" {
+    contract = "C-1FRYVV3"
+    zone = local.zone
     #type = "SECONDARY"
     masters = ["1.2.3.4" , "1.2.3.5"]
     type = "PRIMARY"
     comment =  "This is a test zone"
-    gid     = "64867"
-    signandserve = false
+    group   = "64867"
+    sign_and_serve = false
 }
 
-resource "akamai_dnsv2_record" "a_record" {
-    zone = "${local.zone}"
+resource "akamai_dns_record" "a_record" {
+    zone = local.zone
     name = "akavaiodeveloper.net"
     recordtype =  "A"
     active = true
     ttl =  300
     target = ["10.0.0.2","10.0.0.3"]
-    depends_on =  ["akamai_dnsv2_zone.test_zone"]
+    depends_on =  [akamai_dns_zone.test_zone]
 }
 
-resource "akamai_dnsv2_record" "aaaa_record" {
-    zone = "${local.zone}"
+resource "akamai_dns_record" "aaaa_record" {
+    zone = local.zone
     name = "ipv6record.akavaiodeveloper.net"
     recordtype =  "AAAA"
     active = true
     ttl =  3600
     target = ["2001:0db8::ff00:0042:8329"]
-    depends_on =  ["akamai_dnsv2_zone.test_zone"]
+    depends_on =  [akamai_dns_zone.test_zone]
 }
 
-resource "akamai_dnsv2_record" "afsdb_record" {
-    zone = "${local.zone}"
+resource "akamai_dns_record" "afsdb_record" {
+    zone = local.zone
     name = "afsdb.akavaiodeveloper.net"
     recordtype =  "AFSDB"
     active = true
     ttl =  3600
     subtype = 4
     target = ["example.com"]
-    depends_on =  ["akamai_dnsv2_zone.test_zone"]
+    depends_on =  [akamai_dns_zone.test_zone]
 }
 
 
 
-resource "akamai_dnsv2_record" "cname_record" {
-    zone = "${local.zone}"
+resource "akamai_dns_record" "cname_record" {
+    zone = local.zone
     name = "www.akavaiodeveloper.net"
     recordtype =  "CNAME"
     active = true
     ttl =  300
     target = ["api.akavaiodeveloper.net"]
-    depends_on =  ["akamai_dnsv2_zone.test_zone"]
+    depends_on =  [akamai_dns_zone.test_zone]
 }
 
-resource "akamai_dnsv2_record" "dnskey_record" {
-    zone = "${local.zone}"
+resource "akamai_dns_record" "dnskey_record" {
+    zone = local.zone
     name = "dnskey.akavaiodeveloper.net"
     recordtype =  "DNSKEY"
     active = true
@@ -72,11 +71,11 @@ resource "akamai_dnsv2_record" "dnskey_record" {
     key = "Av//0/goGKPtaa28nQvPoUwVQ ... i/0hC+1CrmQkuuKtQt98WObuv7q8iQ=="
     protocol = 7
     target = ["dnskey.akavaiodeveloper.net"]
-    depends_on =  ["akamai_dnsv2_zone.test_zone"]
+    depends_on =  [akamai_dns_zone.test_zone]
 }
 
-resource "akamai_dnsv2_record" "ds_record" {
-    zone = "${local.zone}"
+resource "akamai_dns_record" "ds_record" {
+    zone = local.zone
     name = "ds.akavaiodeveloper.net"
     recordtype =  "DS"
     active = true
@@ -86,11 +85,11 @@ resource "akamai_dnsv2_record" "ds_record" {
     digest = "909FF0B4DD66F91F56524C4F968D13083BE42380"
     digest_type = 1
     target = ["dnskey.akavaiodeveloper.net"]
-    depends_on =  ["akamai_dnsv2_zone.test_zone"]
+    depends_on =  [akamai_dns_zone.test_zone]
 }
 
-resource "akamai_dnsv2_record" "hinfo_record" {
-    zone = "${local.zone}"
+resource "akamai_dns_record" "hinfo_record" {
+    zone = local.zone
     name = "hinfo.akavaiodeveloper.net"
     recordtype =  "HINFO"
     active = true
@@ -98,34 +97,34 @@ resource "akamai_dnsv2_record" "hinfo_record" {
     hardware = "INTEL-386"
     software = "Unix"
     target = ["hinfo.akavaiodeveloper.net"]
-    depends_on =  ["akamai_dnsv2_zone.test_zone"]
+    depends_on =  [akamai_dns_zone.test_zone]
 }
 
-resource "akamai_dnsv2_record" "loc_record" {
-    zone = "${local.zone}"
+resource "akamai_dns_record" "loc_record" {
+    zone = local.zone
     name = "location.akavaiodeveloper.net"
     recordtype =  "LOC"
     active = true
     ttl =  7200
     target = ["51 30 12.748 N 0 7 39.611 W 0.00m 0.00m 0.00m 0.00m"]
-    depends_on =  ["akamai_dnsv2_zone.test_zone"]
+    depends_on =  [akamai_dns_zone.test_zone]
 }
 
 
-resource "akamai_dnsv2_record" "mx_record" {
+resource "akamai_dns_record" "mx_record" {
     count = 3
-    zone = "${local.zone}"
+    zone = local.zone
     name = "akavaiodeveloper.net"
     recordtype =  "MX"
     active = true
     ttl =  300
     target = ["smtp-${count.index}.akavaiodeveloper.net"]
-    priority = "${count.index*10}"
-    depends_on =  ["akamai_dnsv2_zone.test_zone"]
+    priority = count.index*10
+    depends_on =  [akamai_dns_zone.test_zone]
 }
 
-resource "akamai_dnsv2_record" "naptr_record" {
-    zone = "${local.zone}"
+resource "akamai_dns_record" "naptr_record" {
+    zone = local.zone
     name = "naptrrecord.akavaiodeveloper.net"
     recordtype =  "NAPTR"
     active = true
@@ -137,21 +136,21 @@ resource "akamai_dnsv2_record" "naptr_record" {
     replacement = "."
     service = "SIP+D2U"
     target = ["naptr.akavaiodeveloper.net"]
-    depends_on =  ["akamai_dnsv2_zone.test_zone"]
+    depends_on =  [akamai_dns_zone.test_zone]
 }
 
-resource "akamai_dnsv2_record" "ns_record" {
-    zone = "${local.zone}"
+resource "akamai_dns_record" "ns_record" {
+    zone = local.zone
     name = "ns.akavaiodeveloper.net"
     recordtype =  "NS"
     active = true
     ttl =  300
     target = ["use4.akam.net."]
-    depends_on =  ["akamai_dnsv2_zone.test_zone"]
+    depends_on =  [akamai_dns_zone.test_zone]
 }
 
-resource "akamai_dnsv2_record" "nsec3_record" {
-    zone = "${local.zone}"
+resource "akamai_dns_record" "nsec3_record" {
+    zone = local.zone
     name = "qdeo8lqu4l81uo67oolpo9h0nv9l13dh.akavaiodeveloper.net"
     recordtype =  "NSEC3"
     active = true
@@ -163,11 +162,11 @@ resource "akamai_dnsv2_record" "nsec3_record" {
     salt = "EBD1E0942543A01B"
     type_bitmaps = "CNAME RRSIG"
     target = ["naptr.akavaiodeveloper.net"]
-    depends_on =  ["akamai_dnsv2_zone.test_zone"]
+    depends_on =  [akamai_dns_zone.test_zone]
 }
 
-resource "akamai_dnsv2_record" "nsec3param_record" {
-    zone = "${local.zone}"
+resource "akamai_dns_record" "nsec3param_record" {
+    zone = local.zone
     name = "qnsec3param.akavaiodeveloper.net"
     recordtype =  "NSEC3PARAM"
     active = true
@@ -178,22 +177,22 @@ resource "akamai_dnsv2_record" "nsec3param_record" {
     salt = "EBD1E0942543A01B"
     //salt = "IVBEIMKFGA4TIMRVGQZUCMBRII======"
     target = ["qnsec3param.akavaiodeveloper.net"]
-    depends_on =  ["akamai_dnsv2_zone.test_zone"]
+    depends_on =  [akamai_dns_zone.test_zone]
 }
 
-resource "akamai_dnsv2_record" "ptr_record" {
-    zone = "${local.zone}"
+resource "akamai_dns_record" "ptr_record" {
+    zone = local.zone
     name = "ptr.akavaiodeveloper.net"
     recordtype =  "PTR"
     active = true
     ttl =  300
     target = ["ptr.akavaiodeveloper.net"]
-    depends_on =  ["akamai_dnsv2_zone.test_zone"]
+    depends_on =  [akamai_dns_zone.test_zone]
 }
 
 
-resource "akamai_dnsv2_record" "rp_record" {
-    zone = "${local.zone}"
+resource "akamai_dns_record" "rp_record" {
+    zone = local.zone
     name = "rp.akavaiodeveloper.net"
     recordtype =  "RP"
     active = true
@@ -201,12 +200,12 @@ resource "akamai_dnsv2_record" "rp_record" {
     mailbox = "admin.example.com."
     txt = "txt.example.com"
     target = ["txt.akavaiodeveloper.net"]
-    depends_on =  ["akamai_dnsv2_zone.test_zone"]
+    depends_on =  [akamai_dns_zone.test_zone]
 }
 
 /*
-resource "akamai_dnsv2_record" "rrsig_record" {
-    zone = "${local.zone}"
+resource "akamai_dns_record" "rrsig_record" {
+    zone = local.zone
     name = "rrsig.akavaiodeveloper.net"
     recordtype =  "RRSIG"
     expiration = "20120318104101"
@@ -221,22 +220,22 @@ resource "akamai_dnsv2_record" "rrsig_record" {
     labels = 3
     type_covered = "A"
     target = ["dnskey.akavaiodeveloper.net"]
-    depends_on =  ["akamai_dnsv2_zone.test_zone"]
+    depends_on =  [akamai_dns_zone.test_zone]
 }
 */
 
-resource "akamai_dnsv2_record" "spf_record" {
-    zone = "${local.zone}"
+resource "akamai_dns_record" "spf_record" {
+    zone = local.zone
     name = "spf.akavaiodeveloper.net"
     recordtype =  "PTR"
     active = true
     ttl =  7200
     target = ["v=spf"]
-    depends_on =  ["akamai_dnsv2_zone.test_zone"]
+    depends_on =  [akamai_dns_zone.test_zone]
 }
 
-resource "akamai_dnsv2_record" "srv_record" {
-    zone = "${local.zone}"
+resource "akamai_dns_record" "srv_record" {
+    zone = local.zone
     name = "srv.akavaiodeveloper.net"
     recordtype =  "SRV"
     active = true
@@ -245,11 +244,11 @@ resource "akamai_dnsv2_record" "srv_record" {
     weight  = 0
     port = 522
     target = ["target.akavaiodeveloper.net."]
-    depends_on =  ["akamai_dnsv2_zone.test_zone"]
+    depends_on =  [akamai_dns_zone.test_zone]
 }
 
-resource "akamai_dnsv2_record" "sshfp_record" {
-    zone = "${local.zone}"
+resource "akamai_dns_record" "sshfp_record" {
+    zone = local.zone
     name = "sshfp.akavaiodeveloper.net"
     recordtype =  "SSHFP"
     active = true
@@ -258,40 +257,40 @@ resource "akamai_dnsv2_record" "sshfp_record" {
     fingerprint_type  = 1
     fingerprint = "123456789ABCDEF67890123456789ABCDEF67890"
     target = ["sshfp.akavaiodeveloper.net."]
-    depends_on =  ["akamai_dnsv2_zone.test_zone"]
+    depends_on =  [akamai_dns_zone.test_zone]
 }
 
 
-resource "akamai_dnsv2_record" "txt_record" {
-    zone = "${local.zone}"
+resource "akamai_dns_record" "txt_record" {
+    zone = local.zone
     name = "text.akavaiodeveloper.net"
     recordtype =  "TXT"
     active = true
     ttl =  7200
     target = ["Hello world"]
-    depends_on =  ["akamai_dnsv2_zone.test_zone"]
+    depends_on =  [akamai_dns_zone.test_zone]
 }
 
 /*
-resource "akamai_dnsv2_record" "ns_record" {
-    zone = "${local.zone}"
+resource "akamai_dns_record" "ns_record" {
+    zone = local.zone
     name = "akavaiodeveloper.net"
     recordtype =  "NS"
     active = true
     ttl =  3600
     target = ["use4.akam.net."]
-    depends_on =  ["akamai_dnsv2_zone.test_zone"]
+    depends_on =  [akamai_dns_zone.test_zone]
 }
 */
 
 
 data "akamai_authorities_set" "ns" {
-  contractid = "C-1FRYVV3"
+  contract = "C-1FRYVV3"
 }
 
 
 output "authorities" {
-  value = "${join(",", data.akamai_authorities_set.ns.authorities)}"
+  value = join(",", data.akamai_authorities_set.ns.authorities)
 }
 
 /*
@@ -299,11 +298,11 @@ data "akamai_dns_record_set" "rs" {
   zone = "akavaiodeveloper.net"
   host = "akavaiodeveloper.net"
   record_type = "A"
-  depends_on =  ["akamai_dnsv2_zone.test_zone"]
+  depends_on =  [akamai_dns_zone.test_zone]
 }
 
 output "rs_addrs" {
-  value = "${join(",", data.akamai_dns_record_set.rs.addrs)}"
+  value = join(",", data.akamai_dns_record_set.rs.addrs)
 }
 */
 data "akamai_dns_record_set" "rs1" {
@@ -313,7 +312,7 @@ data "akamai_dns_record_set" "rs1" {
 }
 
 output "rs1_addrs" {
-  value = "${join(",", data.akamai_dns_record_set.rs1.rdata)}"
+  value = join(",", data.akamai_dns_record_set.rs1.rdata)
 }
 
 data "akamai_dns_record_set" "mx" {
@@ -323,5 +322,5 @@ data "akamai_dns_record_set" "mx" {
 }
 
 output "mx_addrs" {
-  value = "${join(",", data.akamai_dns_record_set.mx.rdata)}"
+  value = join(",", data.akamai_dns_record_set.mx.rdata)
 }